refactor: 💡 move utils and update transformDotNotationToNested
