+++ src/components/ThemeEditor/components/DashboardPage/index.ts
@@ -1,5 +1,3 @@
 import DashboardPage from './DashboardPage';
-import { getFontFallback } from './utils';
 
-export { getFontFallback };
 export default DashboardPage;
