--- src/components/ThemeEditor/components/DashboardPage/utils/index.ts
@@ -1,3 +1,0 @@
-import { getFontFallback } from './getFontFallback';
-
-export { getFontFallback };
