+++ src/components/ThemeEditor/index.ts
@@ -1,6 +1,3 @@
 import ThemeEditor from './ThemeEditor';
-import { getFontFallback } from './components/DashboardPage';
-import { mergeSettingsWithFontFallback } from './utils';
 
-export { getFontFallback, mergeSettingsWithFontFallback };
 export default ThemeEditor;
