--- src/components/ThemeEditor/utils/index.ts
@@ -1,4 +1,0 @@
-import { getColorSuggestions } from './getColorSuggestions';
-import { mergeSettingsWithFontFallback } from './mergeSettingsWithFontFallback';
-
-export { getColorSuggestions, mergeSettingsWithFontFallback };
