+++ src/modules/theme/index.ts
@@ -10,7 +10,12 @@ import {
   FONTS,
 } from './constants';
 import { themeSelectors } from './selectors';
-import { extendTheme } from './utils';
+import {
+  extendTheme,
+  getColorSuggestions,
+  mergeSettingsWithFontFallback,
+  getFontFallback,
+} from './utils';
 
 import { ThemeEditorSection, ThemeSettings } from './types';
 
@@ -24,6 +29,9 @@ export {
   themeSagaActions,
   themeSelectors,
   extendTheme,
+  getColorSuggestions,
+  mergeSettingsWithFontFallback,
+  getFontFallback,
   DEFAULT_COLOR_PALETTE,
   CUSTOM_COLOR_PALETTE,
   COLOR_PALETTES,
