+++ src/modules/theme/utils/index.ts
@@ -1,3 +1,11 @@
 import { extendTheme } from './extendTheme';
+import { getColorSuggestions } from './getColorSuggestions';
+import { mergeSettingsWithFontFallback } from './mergeSettingsWithFontFallback';
+import { getFontFallback } from './getFontFallback';
 
-export { extendTheme };
+export {
+  extendTheme,
+  getColorSuggestions,
+  mergeSettingsWithFontFallback,
+  getFontFallback,
+};
